refactoring: Car 에서 Racer 에 해당하는 역할 분리하기
